requirements/base.txt
boto3==1.9.218
celery[tblib]==4.3.0
cffi==1.12.3
channels==2.3.0
channels_redis==2.4.0
chardet==3.0.4
click==7.0
crontab==0.22.6
cryptography==2.7
daphne==2.3.0
dj-database-url==0.5.0
django==2.2.6
django-cors-headers==3.1.1
django-countries-plus==1.2.1
django-filter==2.2.0
django-groups-manager==0.6.2
django-guardian==2.1.0
django-languages-plus==1.0.0
django-mptt==0.10.0
django-nested-inline==0.3.7
django-picklefield==2.0
django-redis==4.10.0
django-rest-auth[with_social]==0.9.5
djangorestframework==3.10.3
distro==1.4.0
drf-dynamic-fields==0.3.1
drf-extensions==0.5.0
drf-proxy-pagination==0.2.0
elasticsearch-dsl==6.3.1
glob2==0.7
jsonfield==2.0.2
lxml==4.4.1
msoffcrypto-tool==4.10.1
natsort==6.0.0
opf-fido==1.3.12
pyfakefs==3.6
python-dateutil==2.8.0
pywin32==225;platform_system=='Windows'
redis==3.3.8
regex==2018.08.29
requests==2.22.0
requests_toolbelt==0.9.1
tenacity==5.1.1
wand==0.5.7
weasyprint==50
